+++ src/main/java/com/plivo/api/models/phonenumbercompliance/PhoneNumberComplianceGetter.java
@@ -1,6 +1,7 @@
 package com.plivo.api.models.phonenumbercompliance;
 
 import com.plivo.api.models.base.Getter;
+import java.util.Map;
 import retrofit2.Call;
 
 public class PhoneNumberComplianceGetter extends Getter<PhoneNumberComplianceGetResponse> {
@@ -18,6 +19,14 @@ public class PhoneNumberComplianceGetter extends Getter<PhoneNumberComplianceGet
     public PhoneNumberComplianceGetter expand(String expand) {
         this.expand = expand;
         return this;
+    }
+
+    @Override
+    protected Map<String, Object> toMap() {
+        Map<String, Object> map = super.toMap();
+        map.remove("id");
+        map.remove("secondaryId");
+        return map;
     }
 
     @Override